Remove unused files

+++ SeafProvider/DocumentPickerViewController.m
@@ -29,7 +29,7 @@
     _conns = SeafGlobal.sharedObject.conns;
     self.tableView.separatorStyle = UITableViewCellSeparatorStyleNone;
     [self.tableView reloadData];
-    Debug("mode: %d", mode);
+    Debug("mode: %lu", (unsigned long)mode);
 }
 
 #pragma mark - Table view data source
